Hide internals of numpy module
A module in Acton correspond to files in the types/ directory etc, which
in turn is the output built from sources in stdlib/src (for all the
built-ins / stdlib). Files are only allowed in this directory with a
direct correspondence to the Acton module.

For example, the numpy module "owns" the stdlib/src/numpy.* "namespace".
Normally there is a .act source file and for modules implemented in C,
there is also a .h and .c file. The .act file and .h both conceptually
contain the interface of the module, just on different levels; in Acton
and on a C level for our intermediate output.

For numpy, we originally chose to have 6 different header files, all
placed in a numpy directory at the root of the acton git repository. When we
recently restructured, we ended up with a single stdlib/src/numpy.h
which then included these 6 header files in stdlib/c_src/numpy/*.h with
some long and weird relative includes. The problem with this is that
includes in stdlib/src/numpy.h need to be resolved at actonc run time,
that is, we must include those libraries with the Acton system. Besides
the logistics inconvenience in including various extra files that follow
no particular structure, it is also bad because it leaks the module
internals to the outside.

To rectify this, the numpy header files in stdlib/c_src/numpy/*.h are
now removed and their content has been inlined into stdlib/src/numpy.h.
It is now the sole header file for the numpy module and contains its
entire interface.

Further, we avoid a similar bleed through of internals by partially
linking the numpy.o file with its dependencies, which in this case is
our own math.o and on linux also libbsd. Outside the compilation of
numpy.o, it is no longer visible what we depend on. As it should!
