+++ Makefile
@@ -1,6 +1,7 @@
 include common.mk
 CHANGELOG_VERSION=$(shell grep '^\#\# \[[0-9]' CHANGELOG.md | sed 's/\#\# \[\([^]]\{1,\}\)].*/\1/' | head -n1)
 
+CFLAGS=-g
 ACTONC=dist/bin/actonc
 
 ifeq ($(shell ls $(ACTONC) >/dev/null 2>&1; echo $$?),0)
@@ -53,10 +54,10 @@ ENV_FILES=$(wildcard builtin/minienv.*)
 BUILTIN_HFILES=$(filter-out $(ENV_FILES),$(wildcard builtin/*.h))
 BUILTIN_CFILES=$(filter-out $(ENV_FILES),$(wildcard builtin/*.c))
 builtin/builtin.o: builtin/builtin.c $(BUILTIN_HFILES) $(BUILTIN_CFILES)
-	cc $(CFLAGS) -Wno-unused-result -g -c -O3 $< -o$@
+	cc $(CFLAGS) -Wno-unused-result -c -O3 $< -o$@
 
 builtin/minienv.o: builtin/minienv.c builtin/minienv.h builtin/builtin.o
-	cc $(CFLAGS) -g -c -O3 $< -o$@
+	cc $(CFLAGS) -c -O3 $< -o$@
 
 # Building the builtin, rts and stdlib is a little tricky as we have to be
 # careful about order. First comes the __builtin__.act file,
@@ -90,13 +91,15 @@ stdlib/out/types/%.h: stdlib/src/%.h
 
 stdlib/out/release/%.o: stdlib/src/%.c
 	@mkdir -p $(dir $@)
-	cc $(CFLAGS) -I. -Istdlib/ -Istdlib/out/ -c $< -o$(subst $,\$,$@)
+	cc $(CFLAGS) -I. -Istdlib/ -Istdlib/out/ -c $< -o$@
 
-NUMPY_HFILES=$(wildcard stdlib/c_src/numpy/*.h)
 NUMPY_CFILES=$(wildcard stdlib/c_src/numpy/*.h)
-stdlib/out/release/numpy.o: stdlib/src/numpy.c stdlib/src/numpy.h stdlib/out/types/math.h $(NUMPY_HFILES) $(NUMPY_CFILES)
+ifeq ($(shell uname -s),Linux)
+NUMPY_CFLAGS+=-lbsd -ldl -lmd
+endif
+stdlib/out/release/numpy.o: stdlib/src/numpy.c stdlib/src/numpy.h stdlib/out/types/math.h $(NUMPY_CFILES) stdlib/out/release/math.o
 	@mkdir -p $(dir $@)
-	cc $(CFLAGS) -Wno-unused-result -I. -Istdlib/out/ -c $< -o$(subst $,\$,$@)
+	cc $(CFLAGS) -Wno-unused-result -r -I. -Istdlib/out/ $< -o$@ $(NUMPY_CFLAGS) stdlib/out/release/math.o
 
 # /lib --------------------------------------------------
 ARCHIVES=lib/libActon.a lib/libActonRTSdebug.a lib/libcomm.a lib/libdb.a lib/libdbclient.a lib/libremote.a lib/libvc.a
@@ -106,11 +109,11 @@ ARCHIVES=lib/libActon.a lib/libActonRTSdebug.a lib/libcomm.a lib/libdb.a lib/lib
 # to form the final libActon (or maybe produce a libActonStdlib and link with?)
 OFILES += builtin/builtin.o builtin/minienv.o $(STDLIB_OFILES) stdlib/out/release/numpy.o rts/empty.o rts/rts.o
 lib/libActon.a: builtin/builtin.o builtin/minienv.o $(STDLIB_OFILES) stdlib/out/release/numpy.o rts/empty.o rts/rts.o
-	ar rcs $@ $(subst $,\$,$^)
+	ar rcs $@ $^
 
 OFILES += rts/rts-debug.o
 lib/libActonRTSdebug.a: rts/rts-debug.o
-	ar rcs $@ $(subst $,\$,$^)
+	ar rcs $@ $^
 
 OFILES += backend/comm.o rts/empty.o
 lib/libcomm.a: backend/comm.o rts/empty.o
@@ -135,13 +138,13 @@ lib/libvc.a: backend/failure_detector/vector_clock.o
 
 # /rts --------------------------------------------------
 rts/rts.o: rts/rts.c rts/rts.h
-	cc $(CFLAGS) -g -Wno-int-to-void-pointer-cast \
+	cc $(CFLAGS) -Wno-int-to-void-pointer-cast \
 		-Wno-unused-result \
 		-pthread \
 		-c -O3 $< -o $@
 
 rts/rts-debug.o: rts/rts.c rts/rts.h
-	cc $(CFLAGS) -DRTS_DEBUG -g -Wno-int-to-void-pointer-cast \
+	cc $(CFLAGS) -DRTS_DEBUG -Wno-int-to-void-pointer-cast \
 		-Wno-unused-result \
 		-pthread \
 		-c -O3 $< -o $@
