--- stdlib/c_src/numpy/init.h
@@ -1,1 +1,0 @@
-void numpy$$__init__();
